Revert "🔨 Consertando erros"
This reverts commit 095a983af7cf1be28babf08c3d6d0bf1a18641db.

+++ app/models/Classificacao.js
@@ -4,7 +4,7 @@ const connection = require('../database/connection')
 class Classificacao extends Model{
     static init(connection){
         super.init({
-            like:DataTypes.INTEGER,
+            like:DataTypes.TINYINT,
             created_at:DataTypes.DATE,
             updated_at:DataTypes.DATE,
             questao_id:DataTypes.BIGINT
